lib: move run-status to agent/lib
Canonical home moves from scripts/lib to agent/lib; scripts-side
importers pick up the #lib/ alias, agent/channels/telegram.ts becomes
a same-tree relative import. Content is unchanged.
