+++ scripts/poller/control.mjs
@@ -1,7 +1,7 @@
 import { createMenu } from "../lib/menu/index.mjs";
 import { botCommands, helpText, tr } from "../lib/i18n.mjs";
 import { continuationTokenForControl } from "../lib/telegram-reset.mjs";
-import { getChatStatus, isRunning } from "../lib/run-status.mjs";
+import { getChatStatus, isRunning } from "#lib/run-status.mjs";
 import { readEnvFresh } from "../lib/env-file.mjs";
 import { formatUsageReport, parseWindow, readEntries, summarize } from "../lib/usage.mjs";
 import { ALLOWED, BOT_USER_ID, DATA_DIR, ENV_PATH, ROOT, log } from "./config.mjs";
