+++ scripts/poller/queue.mjs
@@ -17,7 +17,7 @@ import {
   RUN_STALE_MS,
   setChatStatus,
   setChatStatusIf,
-} from "../lib/run-status.mjs";
+} from "#lib/run-status.mjs";
 import { tr } from "../lib/i18n.mjs";
 import { DATA_DIR, SECRET, RESET_ROUTE, log } from "./config.mjs";
 import { tg } from "./transport.mjs";
