+++ scripts/poller/routing.mjs
@@ -9,7 +9,7 @@ import {
   shouldQueueBusyUpdate,
   TELEGRAM_QUEUE_FATAL_DURABILITY,
 } from "../lib/telegram-queue.mjs";
-import { getChatStatus, isRunning, RUN_STALE_MS, setChatStatusIf } from "../lib/run-status.mjs";
+import { getChatStatus, isRunning, RUN_STALE_MS, setChatStatusIf } from "#lib/run-status.mjs";
 import { tr } from "../lib/i18n.mjs";
 import {
   ACCEPTANCE_ROUTE,
